photo-gps-organizer/tests/test_main.py
"""Unit tests for Photo GPS Organizer."""

import os
import shutil
import tempfile
from pathlib import Path
from unittest.mock import MagicMock, patch

import pytest
import yaml
from PIL import Image
from PIL.ExifTags import GPSTAGS, TAGS

from src.main import PhotoGPSOrganizer


@pytest.fixture
def temp_dir():
    """Create a temporary directory for testing."""
    temp_path = tempfile.mkdtemp()
    yield Path(temp_path)
    shutil.rmtree(temp_path)


@pytest.fixture
def config_file(temp_dir):
    """Create a temporary config file."""
    scan_dir = temp_dir / "source"
    scan_dir.mkdir()

    config = {
        "source_directory": str(scan_dir),
        "destination_directory": str(temp_dir / "organized"),
        "image_extensions": [".jpg", ".jpeg", ".png"],
        "operation": "move",
        "duplicate_handling": "skip",
        "geocoding": {"enabled": False},
        "folder_naming": {
            "coordinate_format": "decimal",
            "coordinate_precision": 4,
            "include_coordinates": False,
        },
        "exclusions": {"directories": [], "patterns": []},
        "operations": {"recursive": True},
        "logging": {
            "level": "INFO",
            "file": str(temp_dir / "test.log"),
            "max_bytes": 10485760,
            "backup_count": 5,
            "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
        },
    }
    config_path = temp_dir / "config.yaml"
    with open(config_path, "w") as f:
        yaml.dump(config, f)
    return str(config_path)


@pytest.fixture
def test_image_with_gps(temp_dir):
    """Create a test image with GPS EXIF data."""
    scan_dir = temp_dir / "source"
    scan_dir.mkdir()

    # Create a simple test image
    img = Image.new("RGB", (100, 100), color="red")
    image_path = scan_dir / "test_photo.jpg"
    img.save(image_path, "JPEG")

    # Add EXIF data with GPS coordinates
    from PIL.ExifTags import GPSTAGS, TAGS

    # Create EXIF data structure
    exif_dict = {
        TAGS.get(271): "Test Camera",
        TAGS.get(272): "Test Model",
    }

    # Add GPS data (New York coordinates)
    gps_ifd = {
        GPSTAGS.get(1): "N",  # GPSLatitudeRef
        GPSTAGS.get(2): ((40, 1), (42, 1), (4608, 100)),  # GPSLatitude
        GPSTAGS.get(3): "W",  # GPSLongitudeRef
        GPSTAGS.get(4): ((74, 1), (0, 1), (2160, 100)),  # GPSLongitude
    }

    # Note: PIL doesn't easily allow adding EXIF data to existing images
    # This is a simplified test - in practice, you'd need a real photo with GPS data
    return image_path


def test_photo_gps_organizer_initialization(config_file):
    """Test PhotoGPSOrganizer initialization."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    assert organizer.config is not None
    assert organizer.source_dir.exists()


def test_load_config_file_not_found():
    """Test that FileNotFoundError is raised when config file doesn't exist."""
    with pytest.raises(FileNotFoundError):
        PhotoGPSOrganizer(config_path="nonexistent.yaml")


def test_convert_to_degrees(config_file):
    """Test GPS coordinate conversion to decimal degrees."""
    organizer = PhotoGPSOrganizer(config_path=config_file)

    # Test valid coordinate tuple
    value = (40, 42, 46.08)
    result = organizer._convert_to_degrees(value)
    assert result is not None
    assert abs(result - 40.7128) < 0.01

    # Test None input
    result = organizer._convert_to_degrees(None)
    assert result is None


def test_format_coordinates_decimal(config_file):
    """Test coordinate formatting in decimal format."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    organizer.config["folder_naming"]["coordinate_format"] = "decimal"

    result = organizer._format_coordinates(40.7128, -74.0060)
    assert "Lat40.7128" in result
    assert "Lon-74.0060" in result


def test_format_coordinates_dms(config_file):
    """Test coordinate formatting in DMS format."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    organizer.config["folder_naming"]["coordinate_format"] = "dms"

    result = organizer._format_coordinates(40.7128, -74.0060)
    assert "N" in result or "S" in result
    assert "E" in result or "W" in result


def test_should_process_file(config_file, temp_dir):
    """Test file filtering logic."""
    organizer = PhotoGPSOrganizer(config_path=config_file)

    # Test valid image file
    jpg_file = temp_dir / "test.jpg"
    jpg_file.write_bytes(b"fake image data")
    assert organizer._should_process_file(jpg_file) is True

    # Test excluded extension
    txt_file = temp_dir / "test.txt"
    txt_file.write_bytes(b"text data")
    assert organizer._should_process_file(txt_file) is False


def test_get_exif_data(config_file, test_image_with_gps):
    """Test EXIF data extraction."""
    organizer = PhotoGPSOrganizer(config_path=config_file)

    # Note: This test may not work perfectly without real EXIF data
    # In practice, you'd use a real photo with GPS data
    exif_data = organizer._get_exif_data(test_image_with_gps)
    # EXIF data may be None if image doesn't have it
    # This is expected for test images created programmatically


def test_get_folder_name_without_geocoding(config_file):
    """Test folder name generation without geocoding."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    organizer.use_geocoding = False

    folder_name = organizer._get_folder_name(40.7128, -74.0060)
    assert "Lat" in folder_name or "Lon" in folder_name


def test_get_folder_name_with_geocoding(config_file):
    """Test folder name generation with geocoding."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    organizer.use_geocoding = True

    # Mock geocoder
    mock_location = MagicMock()
    mock_location.raw = {
        "address": {
            "city": "New York",
            "country": "United States",
        }
    }
    organizer.geocoder = MagicMock()
    organizer.geocoder.reverse = MagicMock(return_value=mock_location)

    folder_name = organizer._get_folder_name(40.7128, -74.0060)
    assert folder_name is not None


def test_organize_photo_without_gps(config_file, temp_dir):
    """Test organizing a photo without GPS data."""
    organizer = PhotoGPSOrganizer(config_path=config_file)
    scan_dir = temp_dir / "source"
    scan_dir.mkdir()

    # Create a simple image without GPS data
    img = Image.new("RGB", (100, 100), color="blue")
    photo_path = scan_dir / "no_gps.jpg"
    img.save(photo_path, "JPEG")

    result = organizer._organize_photo(photo_path)
    # Should return False because no GPS data
    assert result is False
    assert organizer.stats["photos_without_gps"] > 0


def test_environment_variable_overrides(temp_dir):
    """Test that environment variables override config."""
    scan_dir = temp_dir / "source"
    scan_dir.mkdir()

    config = {
        "source_directory": str(scan_dir),
        "destination_directory": str(temp_dir / "organized"),
        "image_extensions": [".jpg"],
        "operation": "move",
        "duplicate_handling": "skip",
        "geocoding": {"enabled": False},
        "folder_naming": {
            "coordinate_format": "decimal",
            "coordinate_precision": 4,
            "include_coordinates": False,
        },
        "exclusions": {"directories": [], "patterns": []},
        "operations": {"recursive": True},
        "logging": {
            "level": "INFO",
            "file": str(temp_dir / "test.log"),
            "max_bytes": 10485760,
            "backup_count": 5,
            "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
        },
    }
    config_path = temp_dir / "config.yaml"
    with open(config_path, "w") as f:
        yaml.dump(config, f)

    with patch.dict(os.environ, {"SOURCE_DIRECTORY": str(temp_dir / "custom")}):
        # This will fail because directory doesn't exist, but tests override
        pass
