photo-gps-organizer/requirements.txt
Pillow==10.1.0  # Image processing and EXIF data extraction
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
geopy==2.4.1  # Reverse geocoding (optional, for location names)
